+++ inlang/source-code/paraglide/paraglide-sveltekit/example/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/paraglide-sveltekit-example
 
+## 0.1.62
+
+### Patch Changes
+
+- Updated dependencies [b5d4a4c]
+  - @inlang/paraglide-sveltekit@0.10.2
+
+## 0.1.61
+
+### Patch Changes
+
+- Updated dependencies [ecef103]
+  - @inlang/paraglide-sveltekit@0.10.1
+
 ## 0.1.60
 
 ### Patch Changes